removed kotlinx.datetime, it was just too annoying and restrictive

+++ src/main/kotlin/de/xehmer/dashboard/widgets/calendar/BaseCalendarWidgetDefinition.kt
@@ -8,5 +8,5 @@ interface BaseCalendarWidgetDefinition : WidgetDefinition {
     val maxEvents: Int
 
     @get:Min(1)
-    val maxDays: Int
+    val maxDays: Long
 }
